The files rename from capitalize
