+++ client/package.json
@@ -25,6 +25,7 @@
     "react-input-range": "^1.3.0",
     "react-map-gl": "^5.2.3",
     "react-mapbox-gl": "^4.8.3",
+    "react-payment-inputs": "^1.1.4",
     "react-redux": "^7.1.3",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
