+++ client/src/components/HeaderFilters/FilteredProducts.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import axios from "axios";
 import { v4 } from "uuid";
 import querystring from "querystring";
-import { ProductItem } from "../ProductsList/productItem";
+import { ProductItem } from "../ProductsList/ProductItem";
 
 import styled from "styled-components";
