+++ client/src/components/ProductsFilterList/FilteredProducts.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import axios from "axios";
 import { v4 } from "uuid";
 import querystring from "query-string";
-import { ProductItem } from "../ProductsList/productItem";
+import { ProductItem } from "../ProductsList/ProductItem";
 import { mediaMobile } from "../../styledComponents/GlobalStyle";
 // import ProductsContainer from "../SliderProducts/ProductsContainer"
