+++ client/src/components/SliderProducts/productSliderView.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Layout } from "../common/Layout";
-import { ProductDetails } from "../ProductDetails/productDetails";
+import { ProductDetails } from "../ProductDetails/ProductDetails";
 import { SliderProducts } from "../SliderProducts/SliderProducts";
 
 export const ProductSliderView = () => {
