+++ client/src/components/StaticPageView/staticPageView.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Layout } from "../common/Layout";
-import { StaticPage } from "../StaticPage/staticPage";
+import { StaticPage } from "../StaticPage/StaticPage";
 
 export const StaticPageView = () => {
   return (
